[scheduler] Add prefer-resize-to-same-host weigher
When an instance is resized, it is rescheduled. So far the
scheduler considers all hosts equally when scheduling an
instance, the current host of the instance is not special.

It is arguably desirable to keep instances on the same host
if possible and keep instance migrations between hosts to a
minimum.

This change addresses this by introducing a prefer-resize-
to-same-host weigher, which weighs the host the instance is
currently running on higher than all others.
This only comes into play if the filters have included the
current host in their output. Notably this change does not
in any way "add" the resources used by the instance to the
available host resources, in short, the current host would
need enough space for "another", resized, instance to be
eligible.

Fixes CCM-13235

Co-Authored-By: Fabian Wiesel <fabian.wiesel@sap.com>
(cherry picked from commit 157d5b258112d63dd03806989bbbe487537c55c0)
(cherry picked from commit 9c8dbaf068f5f6c6d3f8a925d92baefca3a23ac9)
Incorporated-Change-Id: I4b016448a5a905a5d9833aa821daed186d7f1f8a

Change-Id: I3e6598bee888bd815ac8ece9ec1cd4bb0bb93f95