Pass the actual target in limits policy
Currently if target is not passed in context.can(),
it use defauls target which is context.user_id, context.project_id.
These defaults target are not useful as it pass the
context's user_id and project_id only which means we tell
oslo policy to verify the context data with context data.

This commit pass the actual target for limits policies
which is project_id because user_id passed is context.user_id
only.

Partial implement blueprint policy-defaults-refresh

Change-Id: I83523a29d212460b6e4536e7eb0992726c07bd81