Pass the actual target in server external events policy
Currently if target is not passed in context.can(),
it use defauls target which is context.user_id, context.project_id.
These defaults target are not useful as it pass the
context's user_id and project_id only which means we tell
oslo policy to verify the context data with context data.

This commit pass the actual target for server external events policies
which is empty dict because policy rule is system admin scoped.

Adding tests also to show that rule can be override with project
roles.

Partial implement blueprint policy-defaults-refresh

Change-Id: Ia3e0d6a5b8cc5bc8d5c8209a62ffdc771bbf0fa9