Calculate the CPU usage for mixed instance
The mixed instance has two types of CPUs, the shared ones and the
dedicated ones, those CPU usages are tracked in different ways. The
shared CPU is recorded by CPU usage from sharing CPU pool, and the
dedicated CPU is already recorded in 'InstanceNUMACell.cpu_pinning'
when calling 'InstanceNUMACell.pin' method.

This patch enables the usage tracking of the shared CPUs in the
mixed instance.

Part of blueprint use-pcpu-and-vcpu-in-one-instance

Change-Id: I7a31722f1628f47126bb2014555107fffb58aec6
Signed-off-by: Wang Huaqiang <huaqiang.wang@intel.com>