Drop concept of '?validation' parameter
We had planned to introduce a 'validation' parameter to the various
extra spec APIs in change If67f0d924ea372746a6dc440ea7bdc655e4f0bea to
configure a policy for extra spec validation. Following reviews, we've
decided to instead go with a strict policy for all known namespaces and
no validation for any unknown namespaces. Update the tests we had in
place for this.

Part of blueprint flavor-extra-spec-validators

Change-Id: If30990ec1c43177b7d13bd8ee1c5dc481265e47b
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>