Fix new context comparison workaround in base tests class
There is new contexts added in policy test base class that needs
to be added in all tests. Meanwhile we added a workaround not
to assert if that new context is missing.

This commit improve that work around so that we do not assert only
if new context is missing which is from below review comment
- https://review.opendev.org/#/c/716793/5/nova/tests/unit/policies/base.py@138

Partial implement blueprint policy-defaults-refresh

Change-Id: If94865e9fcc92f4c7580315847d3da5d4aebe4a3