Fix server tags policy to be admin_or_owner
server tags API policy is default to admin_or_owner[1] but API
is allowed for everyone.

We can see the test trying with other project context can access the API
- https://review.opendev.org/#/c/717425

This is because API does not pass the server project_id in policy target[2]
and if no target is passed then, policy.py add the default targets which is
nothing but context.project_id (allow for everyone who try to access).

This commit fix this policy by passing the server's project_id in policy
target.

Closes-bug: #1871287
Partial implement blueprint policy-defaults-refresh

[1] https://github.com/openstack/nova/blob/cd16ae25c865f25dbb313976b3d8ef9372db80af/nova/policies/server_tags.py#L27
[2] https://github.com/openstack/nova/blob/cd16ae25c865f25dbb313976b3d8ef9372db80af/nova/api/openstack/compute/server_tags.py#L88

Change-Id: I0718fd16d91b5d208a62f6ced1cb45c1738ab96c