Pass the actual target in server group policy
Currently if target is not passed in context.can(),
it use defauls target which is context.user_id, context.project_id.
These defaults target are not useful as it pass the
context's user_id and project_id only which means we tell
oslo policy to verify the context data with context data.

This commit pass the actual target for server group policies
which is server group project_id for show, delete policy and context
project_id for other policy.

Partial implement blueprint policy-defaults-refresh

Change-Id: Ifdc0828e5989ccdca3bc9139960605c298b0053c