Pass the actual target in keypairs policy
Currently if target is not passed in context.can(),
it use defauls target which is context.user_id, context.project_id.
These defaults target are not useful as it pass the
context's user_id and project_id only which means we tell
oslo policy to verify the context data with context data.

This commit pass the actual target for keypairs policies
which is user _id.

Partial implement blueprint policy-defaults-refresh

Change-Id: Ia0b31a203e138e342f704d595e23af6d669844f1