Pass the actual target in flavor access policy
Currently if target is not passed in context.can(),
it use defauls target which is context.user_id, context.project_id.
These defaults target are not useful as it pass the
context's user_id and project_id only which means we tell
oslo policy to verify the context data with context data.

This commit pass the actual target for flavor access policies
which is empty dict for policy with system scoped and context project_id
for GET which is system and project scoped.

Partial implement blueprint policy-defaults-refresh

Change-Id: I178d619d54d50c9a3d615879e40ab8359f4227b6