Fix server actions to be system and project scoped
This is follow up for review comment:
- https://review.opendev.org/#/c/716130/1/nova/policies/migrate_server.py@46

Let's keep server actions to be both scoped but default them
to SYSTEM-ADMIN/READER and pass the project_id as target. That
way if operator want to enable those policy for project role they
should be able to do.

Also update the defaults for rescue action which found during
audit.

Partial implement blueprint policy-defaults-refresh

Change-Id: If9864ed09d7f604edb8219693dd567638c746456