Test multi create with vGPUs
We had a bug in Rocky where multicreate wasn't working correctly, but given
in Stein we provided Resource Providers for each pGPU, this is fixed now.

NOTE: We have a related bug #1874664 because multicreate doesn't work with
nested Resource Providers.
We could btw. move the regression test to a specific module in the
regressions tests subdirectory.

Change-Id: I8154917ff142987e80dc711e3b2b3965a21f08d0
Related-Bug: #1780225
Related-Bug: #1874664