Don't show upgr note for policy validation in V
We merged Id9cd65877e53577bff22e408ca07bbeec4407f6e which included a reno
file after we branched Ussuri so we had to backport it to Ussuri but since
it was merged after RC1, the relnote would appear as an Upgrade section for
the Victoria release, which is not something we wanted (but rather for a
T->U upgrade).

Adding the ignore-notes section in order to avoid it.

Change-Id: If31933c5ff20167bf24fbad6f8746b1a9a6a36e5