libvirt: apply mixed instance CPU policy
For an instance with a 'mixed' CPU allocation policy, the
dedicated CPUs and the shared CPUs are mixed in one instance.

From the host CPU using perspective, dedicated CPU exclusively
occupies one host CPU but may offer some CPU time slots to instance
emulator if the emulator thread policy is not 'ISOLATE'. While the
shared CPU is sharing host CPU with other instance CPUs, the
shared CPU floats over a list of host CPUs of a NUMA cell if the
instance is NUMA-aware. The emulator thread is running on the
reserved host CPU if the emulator thread policy is 'ISOLATE',
if the policy is not 'ISOLATE', it will float on host CPUs listing
in '[compute] cpu_shared_set' if this configuration is provided
and the emulator thread policy is explicitly set as 'SHARE'.
otherwise, the emulator thread is floating on all non-realtime CPUs
that the instance is used.

Later, we'll introduce the 'mixed' instance mixing with realtime
CPUs and shared CPUs.

Change-Id: I0ea3a0a1342ef53ed891c145561eb7402f553a8b
Signed-off-by: Wang Huaqiang <huaqiang.wang@intel.com>