Make quotas respect instance_list_per_project_cells
This option was introduced in order to limit queries to cells which only
belong to the project we're interacting with.  However, The nova quota
code was not respecting it and therefore it was always trying to get the
cells assigned to a project even with that option disabled.

This patch makes the quota code respect that option and adds testing to
ensure that enabling the option does make sure it doesn't search all
cells but only the ones for this project.

Closes-Bug: #1878979
Change-Id: I2e0d48e799e70d550f912ad8a424c86df3ade3a2
(cherry picked from commit ab16946885e68ccdef7222a71cc0ad6f92b10de7)