Add link to PDF document
Switch to openstackdocstheme 2.2.0 that can link to PDF document,
enable this with setting openstackdocs_pdf_link.

Note that the link to the published document only works on 
docs.openstack.org where the PDF file is placed in the top-level html
directory. The site-preview places the PDF in a pdf directory.

Depends-On: https://review.opendev.org/728938
Change-Id: I29dddfa06183ecadb8f23feb2ad2e3c1c5a233cc