Raise 409 when removing security group from instance
currently when attempting to remove a security group from instance, nova
api returns 500 when there are more than one secuity groups with such
name in the project.

This patch makes it return 409 Conflict, using the same logic as already
present when adding a security group to the instance in such scenario of
duplicate names for security groups.

Change-Id: I28ad999c49d4e11add54405b89185b3f56f165b5
Closes-Bug: 1889655